requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anticaptchaofficial==1.0.61
anyio==4.6.2.post1
async-timeout==4.0.3
atlastk==0.13.3
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.2
certifi==2024.8.30
chardet==4.0.0
charset-normalizer==3.4.0
curlify==3.0.0
exceptiongroup==1.2.2
facebook_business==23.0.1
FacebookAdsWindsorApi==1.0.6
frozenlist==1.5.0
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==2.10
iso8601==2.1.0
lxml==5.3.0
magic-filter==1.0.12
multidict==6.1.0
oauthlib==3.3.1
propcache==0.2.0
proto-plus==1.26.1
protobuf==6.31.1
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pyotp==2.9.0
pyparsing==3.2.3
pypika-tortoise==0.2.1
python-anticaptcha==1.0.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.25.1
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.2
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tortoise==0.1.1
tortoise-orm==0.21.7
TwoCaptcha==0.0.1
typing_extensions==4.12.2
uritemplate==4.2.0
urllib3==1.26.20
yarl==1.17.1
